wip: export logic to new geocoding class
